Update examples to work with latest version
This includes sub-optimal repitition of transition event handlers: we must
bind the same handler to both the `enter:transition` event *and* the
`update:transition` event. Ideally, we would bind it once to the
`merge:transition` event, but making this selection cancels
previously-scheduled transitions on the `enter` and `update` selections...

This also highlights a problem with the current mixin API: handlers for
lifecycle events in "mixin" charts cannot access the chart into which they
have been mixed. In other words, layers in these mix in charts can only
access the chart they have been defined within...